+++ Radio3.0@claudiux/files/Radio3.0@claudiux/CHANGELOG.md
@@ -1,3 +1,16 @@
+### v1.11.1~20231128
+  * Moves the "Do not check dependencies" option from the Settings Behavior tab to the context menu. This avoids error messages.
+
+### v1.11.0~20231128
+  * Fixes dependency issues.
+  * Option "Do not check about dependencies" works fine now.
+
+### v1.10.1~20231126
+  * Remove libmpv1 or libmpv2 dependency for LM and Debian.
+
+### v1.10.0~20231126
+  * Added the ability to swap artist and song title for certain radio stations.
+
 ### v1.9.1~20231115
   * Fixes #5097 and removes mpv-mpris for versions of distros that do not have the right package.